chore: cleanup and remove loads of old code
