+++ src/datasets/spectral-dataset.jl
