+++ src/fitting/statistics.jl
@@ -5,12 +5,12 @@ measure(::ChiSquared, y, ŷ, σ²) = sum(@.((y - ŷ)^2 / σ²))
 
 struct Cash <: AbstractStatistic end
 
-function wrap_objective(stat::AbstractStatistic, f, config::FittingConfig)
-    function _objective(u, x)
-        ŷ = f(x, u)
-        measure(stat, config.y, ŷ, config.variance)
+function _f_wrap_objective(stat::AbstractStatistic, config::FittingConfig)
+    f = _f_objective(config)
+    function _objective(parameters, domain)
+        ŷ = f(domain, parameters)
+        measure(stat, config.objective, ŷ, config.variance)
     end
 end
-
 
 export ChiSquared, Cash
