+++ src/plotting-recipes.jl
@@ -1,24 +1,24 @@
 using RecipesBase
 
-@recipe function _plotting_func(d::SpectralDataset)
-    seriestype --> :scatter
-    markersize --> 0.5
-    markershape --> :none
-    (rate, rateerror) = (get_rate(d), get_rate_variance(d))
-    yerr --> rateerror
-    xerr --> get_bin_widths(d) ./ 2
-    markerstrokecolor --> :auto
-    xscale --> :log10
-    yscale --> :log10
-    yticks --> ([0.01, 0.1, 1, 10, 100], [0.01, 0.1, 1, 10, 100])
-    xticks --> ([1e-1, 1, 2, 5, 10, 20, 50, 100], [1e-1, 1, 2, 5, 10, 20, 50, 100])
-    xlabel --> "Energy (keV)"
-    ylabel --> "counts s⁻¹ keV⁻¹"
-    label --> observation_id(d)
-    minorgrid --> true
-    x = ((d.bins_low.+d.bins_high)./2)[d.mask]
-    (x, rate)
-end
+# @recipe function _plotting_func(d::SpectralDataset)
+#     seriestype --> :scatter
+#     markersize --> 0.5
+#     markershape --> :none
+#     (rate, rateerror) = (get_rate(d), get_rate_variance(d))
+#     yerr --> rateerror
+#     xerr --> get_bin_widths(d) ./ 2
+#     markerstrokecolor --> :auto
+#     xscale --> :log10
+#     yscale --> :log10
+#     yticks --> ([0.01, 0.1, 1, 10, 100], [0.01, 0.1, 1, 10, 100])
+#     xticks --> ([1e-1, 1, 2, 5, 10, 20, 50, 100], [1e-1, 1, 2, 5, 10, 20, 50, 100])
+#     xlabel --> "Energy (keV)"
+#     ylabel --> "counts s⁻¹ keV⁻¹"
+#     label --> observation_id(d)
+#     minorgrid --> true
+#     x = ((d.bins_low.+d.bins_high)./2)[d.mask]
+#     (x, rate)
+# end
 
 @recipe function _plotting_func(d::SimpleDataset)
     seriestype --> :scatter
